requirements/api.txt
-r common.txt

djangorestframework==3.3.2
